+++ src/chattool/setup/alias.py
@@ -10,13 +10,12 @@ logger = setup_logger("setup_alias")
 
 ALIAS_MAP = {
     "chatenv": "chattool env",
-    "chatskills": "chattool skill",
+    "chatskill": "chattool skill",
     "chatdns": "chattool dns",
     "chatgh": "chattool gh",
     "chatcc": "chattool cc",
     "chatcli": "chattool client",
     "chatpypi": "chattool pypi",
-    "chatdns": "chattool dns",
     "chatup": "chattool setup",
     "chatlark": "chattool lark",
     "chatimg": "chattool image",